requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
coverage==7.6.3
datetype==2024.2.28
flake8==7.0.0
flake8-html==0.4.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
locust==2.31.8
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.1.0
npm==0.1.1
optional-django==0.1.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psutil==6.0.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.3.3
pytest-flask==1.3.0
pytest-mock==3.14.0
python-slugify==8.0.4
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
text-unidecode==1.3
urllib3==2.2.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
